+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
-## <a id="v1.6.10"></a>[v1.6.10] - 2022-07-24
+## <a id="v1.6.11"></a>[v1.6.11] - 2022-07-26
 
 ### Added
 
@@ -22,6 +22,8 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - analysis service stuck when storage has ambiguous directories
+
+## <a id="v1.6.10"></a>[v1.6.10] - 2022-07-24 [YANKED]
 
 ## <a id="v1.6.9"></a>[v1.6.9] - 2022-06-18